refactor: create middleware package
• Relocated middleware class from server.py to middleware/bindings.py
• Renamed DynamicToolInjectionMiddleware to BindingsMiddleware
• Updated server.py to import from new location
• Fixed test imports and class references
• Removed unused middleware imports from server.py
